refactor: remove large amounts of copypasta from platform responses
