+++ lib/methods/platform/getIdentitiesByPublicKeyHashes/GetIdentitiesByPublicKeyHashesResponse.js
@@ -1,7 +1,4 @@
 const AbstractResponse = require('../response/AbstractResponse');
-const Metadata = require('../response/Metadata');
-const InvalidResponseError = require('../response/errors/InvalidResponseError');
-const createProofFromRawProof = require('../response/createProofFromRawProof');
 
 class GetIdentitiesByPublicKeyHashesResponse extends AbstractResponse {
   /**
@@ -27,23 +24,12 @@ class GetIdentitiesByPublicKeyHashesResponse extends AbstractResponse {
    * @returns {GetIdentitiesByPublicKeyHashesResponse}
    */
   static createFromProto(proto) {
-    const metadata = proto.getMetadata();
-
-    if (metadata === undefined) {
-      throw new InvalidResponseError('Metadata is not defined');
-    }
-
-    const rawProof = proto.getProof();
-
-    let proof;
-    if (rawProof) {
-      proof = createProofFromRawProof(rawProof);
-    }
+    const { metadata, proof } = AbstractResponse.createMetadataAndProofFromResponse(proto);
 
     return new GetIdentitiesByPublicKeyHashesResponse(
       proto.getIdentitiesList()
         .map((identity) => (identity.length > 0 ? Buffer.from(identity) : null)),
-      new Metadata(metadata.toObject()),
+      metadata,
       proof,
     );
   }
