+++ lib/methods/platform/waitForStateTransitionResult/WaitForStateTransitionResultResponse.js
@@ -2,8 +2,8 @@ const cbor = require('cbor');
 
 const AbstractResponse = require('../response/AbstractResponse');
 const Metadata = require('../response/Metadata');
+const Proof = require('../response/Proof');
 const ErrorResult = require('./ErrorResult');
-const createProofFromRawProof = require('../response/createProofFromRawProof');
 
 class WaitForStateTransitionResultResponse extends AbstractResponse {
   /**
@@ -33,7 +33,7 @@ class WaitForStateTransitionResultResponse extends AbstractResponse {
     let proof;
 
     if (proto.getProof()) {
-      proof = createProofFromRawProof(proto.getProof());
+      proof = Proof.createFromProto(proto.getProof());
     }
 
     if (proto.getError()) {
